refac: refactor for service package by domain
