+++ src/main/java/coLaon/ClaonBack/post/service/PostLikeService.java
@@ -1,0 +1,112 @@
+package coLaon.ClaonBack.post.service;
+
+import coLaon.ClaonBack.common.domain.Pagination;
+import coLaon.ClaonBack.common.domain.PaginationFactory;
+import coLaon.ClaonBack.common.exception.BadRequestException;
+import coLaon.ClaonBack.common.exception.ErrorCode;
+import coLaon.ClaonBack.common.exception.UnauthorizedException;
+import coLaon.ClaonBack.post.domain.Post;
+import coLaon.ClaonBack.post.domain.PostLike;
+import coLaon.ClaonBack.post.dto.LikeFindResponseDto;
+import coLaon.ClaonBack.post.dto.LikeResponseDto;
+import coLaon.ClaonBack.post.repository.PostLikeRepository;
+import coLaon.ClaonBack.post.repository.PostRepository;
+import coLaon.ClaonBack.user.domain.User;
+import coLaon.ClaonBack.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class PostLikeService {
+    private final UserRepository userRepository;
+    private final PostRepository postRepository;
+    private final PostLikeRepository postLikeRepository;
+    private final PaginationFactory paginationFactory;
+
+    @Transactional
+    public LikeResponseDto createLike(String userId, String postId) {
+        User liker = userRepository.findById(userId).orElseThrow(
+                () -> new UnauthorizedException(
+                        ErrorCode.USER_DOES_NOT_EXIST,
+                        "이용자를 찾을 수 없습니다."
+                )
+        );
+
+        Post post = postRepository.findByIdAndIsDeletedFalse(postId).orElseThrow(
+                () -> new BadRequestException(
+                        ErrorCode.ROW_DOES_NOT_EXIST,
+                        "게시글을 찾을 수 없습니다."
+                )
+        );
+
+        postLikeRepository.findByLikerAndPost(liker, post).ifPresent(
+                like -> {
+                    throw new BadRequestException(
+                            ErrorCode.ROW_ALREADY_EXIST,
+                            "이미 좋아요 한 게시글입니다."
+                    );
+                }
+        );
+
+        return LikeResponseDto.from(
+                postLikeRepository.save(PostLike.of(liker, post)),
+                postLikeRepository.countByPost(post)
+        );
+    }
+
+    @Transactional
+    public LikeResponseDto deleteLike(String userId, String postId) {
+        User liker = userRepository.findById(userId).orElseThrow(
+                () -> new BadRequestException(
+                        ErrorCode.ROW_DOES_NOT_EXIST,
+                        "이용자를 찾을 수 없습니다."
+                )
+        );
+
+        Post post = postRepository.findByIdAndIsDeletedFalse(postId).orElseThrow(
+                () -> new BadRequestException(
+                        ErrorCode.ROW_DOES_NOT_EXIST,
+                        "게시글을 찾을 수 없습니다."
+                )
+        );
+
+        PostLike like = postLikeRepository.findByLikerAndPost(liker, post).orElseThrow(
+                () -> new BadRequestException(
+                        ErrorCode.ROW_DOES_NOT_EXIST,
+                        "좋아요 하지 않은 게시글입니다."
+                )
+        );
+
+        postLikeRepository.delete(like);
+
+        return LikeResponseDto.from(
+                like,
+                postLikeRepository.countByPost(like.getPost())
+        );
+    }
+
+    @Transactional(readOnly = true)
+    public Pagination<LikeFindResponseDto> findLikeByPost(String userId, String postId, Pageable pageable) {
+        userRepository.findById(userId).orElseThrow(
+                () -> new UnauthorizedException(
+                        ErrorCode.USER_DOES_NOT_EXIST,
+                        "이용자를 찾을 수 없습니다."
+                )
+        );
+
+        Post post = postRepository.findById(postId).orElseThrow(
+                () -> new BadRequestException(
+                        ErrorCode.ROW_DOES_NOT_EXIST,
+                        "게시글을 찾을 수 없습니다."
+                )
+        );
+
+        return this.paginationFactory.create(
+                postLikeRepository.findAllByPost(post, pageable)
+                        .map(LikeFindResponseDto::from)
+        );
+    }
+}
