+++ src/test/java/coLaon/ClaonBack/service/PostServiceTest.java
@@ -8,8 +8,6 @@ import coLaon.ClaonBack.center.domain.OperatingTime;
 import coLaon.ClaonBack.center.domain.SectorInfo;
 import coLaon.ClaonBack.center.repository.CenterRepository;
 import coLaon.ClaonBack.center.repository.HoldInfoRepository;
-import coLaon.ClaonBack.common.domain.Pagination;
-import coLaon.ClaonBack.common.domain.PaginationFactory;
 import coLaon.ClaonBack.common.exception.ErrorCode;
 import coLaon.ClaonBack.common.exception.UnauthorizedException;
 import coLaon.ClaonBack.post.domain.ClimbingHistory;
@@ -18,16 +16,13 @@ import coLaon.ClaonBack.post.dto.PostContentsDto;
 import coLaon.ClaonBack.post.dto.PostCreateRequestDto;
 import coLaon.ClaonBack.post.dto.PostDetailResponseDto;
 import coLaon.ClaonBack.post.dto.PostResponseDto;
-import coLaon.ClaonBack.post.dto.LikeResponseDto;
-import coLaon.ClaonBack.post.dto.LikeFindResponseDto;
 import coLaon.ClaonBack.post.repository.ClimbingHistoryRepository;
+import coLaon.ClaonBack.post.repository.PostLikeRepository;
 import coLaon.ClaonBack.post.service.PostService;
 import coLaon.ClaonBack.post.domain.Post;
 import coLaon.ClaonBack.post.domain.PostContents;
-import coLaon.ClaonBack.post.domain.PostLike;
 import coLaon.ClaonBack.common.exception.BadRequestException;
 import coLaon.ClaonBack.post.repository.PostContentsRepository;
-import coLaon.ClaonBack.post.repository.PostLikeRepository;
 import coLaon.ClaonBack.post.repository.PostRepository;
 import coLaon.ClaonBack.user.domain.BlockUser;
 import coLaon.ClaonBack.user.domain.User;
@@ -41,12 +36,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -56,7 +46,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.tuple;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mockStatic;
 
@@ -78,13 +67,10 @@ public class PostServiceTest {
     ClimbingHistoryRepository climbingHistoryRepository;
     @Mock
     BlockUserRepository blockUserRepository;
-    @Spy
-    PaginationFactory paginationFactory = new PaginationFactory();
 
     @InjectMocks
     PostService postService;
 
-    private PostLike postLike, postLike2;
     private User user, user2, blockedUser, privateUser;
     private Post post, post2, blockedPost, privatePost;
     private PostContents postContents, postContents2;
@@ -189,18 +175,6 @@ public class PostServiceTest {
                 "test2.com/test.png"
         );
 
-        this.postLike = PostLike.of(
-                "testPostLikeId",
-                user,
-                post
-        );
-
-        this.postLike2 = PostLike.of(
-                "testPostLikeId2",
-                user2,
-                post
-        );
-
         this.holdInfo1 = HoldInfo.of(
                 "holdId1",
                 "holdName1",
@@ -262,6 +236,7 @@ public class PostServiceTest {
         given(this.postRepository.findByIdAndIsDeletedFalse("testPostId2")).willReturn(Optional.of(post2));
         given(this.userRepository.findById("testUserId")).willReturn(Optional.of(user));
         given(this.blockUserRepository.findBlock("testUserId", user2.getId())).willReturn(Optional.empty());
+        given(this.postLikeRepository.countByPost(post2)).willReturn(2);
 
         // when
         PostDetailResponseDto postResponseDto = this.postService.findPost("testUserId", "testPostId2");
@@ -440,73 +415,5 @@ public class PostServiceTest {
         assertThat(ex)
                 .extracting("errorCode", "message")
                 .contains(ErrorCode.NOT_ACCESSIBLE, "접근 권한이 없습니다.");
-    }
-
-    @Test
-    @DisplayName("Success case for create like")
-    void successCreateLike() {
-        try (MockedStatic<PostLike> mockedPostLike = mockStatic(PostLike.class)) {
-            // given
-            given(this.userRepository.findById("testUserId")).willReturn(Optional.of(user));
-            given(this.postRepository.findByIdAndIsDeletedFalse("testPostId")).willReturn(Optional.of(post));
-            given(this.postLikeRepository.findByLikerAndPost(user, post)).willReturn(Optional.empty());
-
-            mockedPostLike.when(() -> PostLike.of(user, post)).thenReturn(postLike);
-            given(this.postLikeRepository.countByPost(post)).willReturn(1);
-
-            given(this.postLikeRepository.save(this.postLike)).willReturn(postLike);
-
-            // when
-            LikeResponseDto likeResponseDto = this.postService.createLike("testUserId", "testPostId");
-
-            // then
-            assertThat(likeResponseDto)
-                    .isNotNull()
-                    .extracting("postId", "likeCount")
-                    .contains("testPostId", 1);
-        }
-    }
-
-    @Test
-    @DisplayName("Success case for delete like")
-    void successDeleteLike() {
-        // given
-        given(this.userRepository.findById("testUserId")).willReturn(Optional.of(user));
-        given(this.postRepository.findByIdAndIsDeletedFalse("testPostId")).willReturn(Optional.of(post));
-        given(this.postLikeRepository.findByLikerAndPost(user, post)).willReturn(Optional.of(postLike));
-
-        // when
-        LikeResponseDto likeResponseDto = this.postService.deleteLike("testUserId", "testPostId");
-
-        // then
-        assertThat(likeResponseDto)
-                .isNotNull()
-                .extracting("postId", "likeCount")
-                .contains("testPostId", 0);
-    }
-
-    @Test
-    @DisplayName("Success case for find likes")
-    void successFindLikes() {
-        // given
-        Pageable pageable = PageRequest.of(0, 2);
-        given(this.userRepository.findById("testUserId")).willReturn(Optional.of(user));
-        given(this.postRepository.findById("testPostId")).willReturn(Optional.of(post));
-
-        Page<PostLike> postLikes = new PageImpl<>(List.of(postLike, postLike2), pageable, 2);
-
-        given(this.postLikeRepository.findAllByPost(post, pageable)).willReturn(postLikes);
-
-        // when
-        Pagination<LikeFindResponseDto> likeFindResponseDto = this.postService.findLikeByPost("testUserId", "testPostId", pageable);
-
-        // then
-        assertThat(likeFindResponseDto.getResults())
-                .isNotNull()
-                .extracting(LikeFindResponseDto::getPostId, LikeFindResponseDto::getLikerNickname)
-                .contains(
-                        tuple("testPostId", postLike.getLiker().getNickname()),
-                        tuple("testPostId", postLike2.getLiker().getNickname())
-                );
     }
 }
